backend/foodgram/requirements.txt
Django==2.2.19
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
Pillow==8.3.1
python-dotenv==0.19.2
pytz==2022.7.1
sqlparse==0.4.3
